DeviceN & psdcmykog devices - improve name parsing in SeparationOrder
No bug for this, revealed during release testing. The psdcmykog device
with the test file /tests_private/ps/ps3cet/29-07g.ps.

Because of commit 92ff66cc7a3884c99a2508d1b6ced076a4526d13 we started
throwing errors with this device and file, because the file tries to
set a DeviceN ProcessColorModeal and then set colorants which are not
available to the psdcmykog device.

At least some of the errors are correct, because the psdcmykog device
does not permit additional spot plates, but it should permit the
CMYK case and that was throwing an error.

This commit changes the number of colorants in the psdcmykog device to
6, which I hope is the correct thing to do (4 is certainly wrong) and
moves the check against the number of permitted components to being
done after checking whether a name in the SeparationOrder is already
present in the process colorants list.