DeviceN devices - yet again
There is considerable confusion in the DeviceN code about the meaning of
'std_colorant_names' and 'num_std_colorant_names'. For example the
CMYKOG device has std_coloranr_names include Cyan, Magenta, Yellow,
Black, Artifex Orange and Artifex Green. But num_std_colorant_names is
4.

It appears that num_std_colorant_names is used to indicate, in effect,
the number of inks in the base colour model, 1, 3 or 4 for Gray, RGB
or CMYK. There is code which uses it as such as well.

In commit 33daeee2311a926da3e001d17ae0f5504b6272e5 I altered the
number of components in the CMYKOG device to be 6, because that altered
the num_std_colorant_names to match the std_colorant_names list, which
is what I expected it should be. This turns out to be an error, so
this commit reverts that as a first step.

But we need the number of 'pre-defined' colorants in devn_put_params()
so that we can check that number of predefined inks + number of spots
doesn't exceed the maximum number of components.

So I've added a function to count the number of inks in the list, and
then used that instead of num_std_colorant_names.