Optimise clip_fill_path.
A 'fill_path' operation does 'fill this path clipped by this
clip path'.

When playing a 'fill_path' operation back through an existing
clip path (such as when playing a pattern clist back through a
clip path), we end up calling the 'clip_fill_path' operator.

Currently, this is implemented by walking through the rectangles
that make up the "outer" clip path, looking for rectangles that
intersect with the bbox of the path.

For each rectangle we find, we make a new clip path from
the intersection of the "inner" clip path and the rectangle, and
"fill_path" with that new clip path. Then we discard that new
clip path.

This means a lot of intersection operations, which are slow.

In this new code, we modify this behaviour. We still walk through
the rectangles that make up the "outer" clip path, looking for
rectangles that intersect with the bbox of the path.

For each rectangle we find, we then walk through the rectangles
that make up the "inner" clip path that intersect with that.

For each of these nested rectangles, we can form a simple clip
path based upon that rectangle, and call fill_path with that.

This avoids the expensive intersection operation.