pdfwrite - fix 32 bit build and XRef streams
Bug #707610 "gs10030w32.exe 32bit version "-sDEVICE=pdfwrite" creates damaged pdf file"

The PRId64 macro doesn't work well when the data type is a 'long' on
Windows, and the word size is 32 bits.

A better way to fix this is to promote the structure members to be
64-bit using the int64_t data type, but that's too invasive a change
while we are in the middle of a release.

For now fix the actual problem by casting the data types to the higher
type.