pdfwrite - remove most uses of 'long' data type
Prompted by bug 707610 where the use of a 'long' data type caused an
error with the PRId64 macro.

On Windows, at least, the 'long' data type is a different size on 32
and 64-bit builds. This potentially could cause quite different results
with pdfwrite. So here we try to reduce the usage of long, and replace
it with a data type which is the same on all platforms.

In particular 'id' members were often defined as long, which meant the
number of object ids would be smaller on 32-bit builds and there's no
real reason for that. I think we are more consistent now.