pdfwrite - Fix an error introduced by changing long
I altered the use of 'long' in pdfwrite to avoid problems with the
PRId64 macro, but I didn't realise that pprintld1 didn't implement it.

For some reason when I tested this commit no differences showed up,
but when the commit test run executed it did show differences. I'm
unable to reproduce them on Linux or Windows but this problem did
show up as a warning so I'm fixing this in the hope that the differences
will disappear.