Update scripts to avoid warnings with new features
If the Compatibility Level is less than 1.5 then the default value
of WriteXRefStm and WriteObjStms is invalid, and we emit a warning to
the user.

But the shell scripts mostly default to 1.4 or less, meaning we always
emit a warning.

Modify the scripts which involve a COmpatibilityLevel less than 1.5 so
that the defaults are overriden to avoid the warnings.