pdfwrite - preserve original non-embedded font Subtype
Extra work for Bug #707578. Building on Chris's commit
7d3e88528a466ca81df4196db5322da9aab16667 and following on from the
work in commits 99b48f1992844e30ab4b05dbb2764ba99f7b66f4 and
fd14cd681da01783ae8df0fc71e653cc2c218cec.

We now do not embed a substittue font in a PDF otuptu file if the input
was PDF, and the input PDF file did not have the font embedded in it.
We now additionally preserve the Subtype of the original font, this is
useful for mapping glyphs correctly (or at least, consistently!) with
the original PDF file.

This shows a difference with the file for bug 707578; if we render the
new PDF file with a FONTPATH or fontmap.GS which includes TrueType
substitutions for the non-embedded fonts. Previously we altered the
font /Subtype to type 1 (our substituted font) which rendered the wrong
glyphs if we substituted a TrueType font when rendering. Now we
draw the correct glyphs, the same glyphs as when rendering the original
file.