pdfwrite - drop unused named objects from the xref & fix xrefs
Noticed while working on something else entirely. If we get an error
while handling an annotation after we've created a named dictionary we
freed the dictionary, but that still left a slot allocated for it when
we wrote out the xref.

This has two undesirable consequences; firstly we are writing xref
entries we don't need to which makes th file larger, and secondly the
offset we were writing was past the end of the file. This doesn't
actually matter because nothing should reference these objects from the
xref (the objects don't exist) but it's technically wrong.

So if an annotation fails, and we've allocated a named dictionary,
remove the named dictionary from the xref.

This then highlighted a different problem with writing xref sections
with both normal xref tables and XRef xtreams. We weren't properly
skipping empty sections.