Revisit the DeviceN code, again....
This commit rewrites the separation handling. We now check each ink
in /SeparationColorNames against the devices 'standard' colorant names
and against any separations already defined.

If a coloran name matches either of these it is not new.

Having established the number of new colorants, we check to see that the
number of standard colorants, plus the number of existing spots, plus
the number of new spots does not exceed the maximum number of components
for the device. If it does we error.

We then proceed to repeat the naem testing, but this time we add each
new colorant to the separration.names array. We know they must fit
because we checked previously.

This now prevents us adding the same colorant multiple times and avoids
overrunning the array.

This does show differences with the cluster but these appear to be due
to the colorant order changing, though I can't see why it would and
doesn't when run locally.