pdfwrite - more improvements for mesh shadings
Bug #707655 "No output visible with pdfwriter"

commit 3d6e3acbcda79a0096cd1ad73c7b9b1101a43187 to fix bug #06852
unfortunately led to this regression. The problem is related to the
scaling of co-ordinates performed when a mesh shading has vertex
co-ordinates which are too large for Acrobat 5 (PDF 1.4).

Acrobat 5 has a limit of =/-32767 on real numbers, which applies to the
/Decode array of a mesh shading. Because this is an Acrobat limit, not a
limit of the PDF specification we would ordinarily ignore it, but the
PDF/A-1 specification chose to carry the Acrobat limitations into the
PDF/A spec.

This commit fixes the problem by correctly scaling the co-ordinate
values to the Decode array values when outputting to PDF 1.4 or less,
even if no actual co-ordinte scaling is required, and modifying the
/Matrix of the Pattern to scale up the Decode values to the real
co-ordinates. Crucially we must not scale the Tx and Ty values of the
CTM.