pdfwrite - disable XRefStm when writing Linearized files (FastWebView)
Bug #707653 "-dFastWebView=true does not work anymore"

I'd disabled ObjStm usage, but I missed that we need to disable XRefStm
use as well, when writing linearised files.

While I'm messing with this anyway; alter the handling so that we
detect whether the user has specifically selected WriteXRefStm and
WriteObjStms or not. If they have not been specifically requested and
are not compatible, silently disable them. If they have been
specifically enabled, and are not compatible with other settings, warn
the user and disable them anyway.