PDF interpreter - remove a spurious dictionary
Bug #707660 " Crash rendering landscape PDF to PNG"

The title s wrong, it isn't a crash at all. The problem is that the
reporter is using PDFFitPage with a pahe which is landscape and does
not have a /Rotate key of any kind.

This inadvertantly leaves a copied dictionary on the stack, which then
leads to a typecheck error.

Remove the dictionary.