pdfwrite - properly handle error condition with shading rendering
No bug report, this is from a file supplied by a customer. The PDF file
has numerous (> 2500) shadings, defined in CMYK space which are
degenerate but drawn radically off the media. When run through pdfwrite
with ColorConversionStrategy=/RGB this crashes.

The reason is that the masked image converter which is used to render the
shadings into an image doesn't like the fact that the content is
entirely off the media and returns an error. The pdfwrite code spots
the error, but there's a logic mistake in the flow of control which
tries to remove the masked image converter, which is not possible
because we failed to install it, and leads to a crash.

This commit fixes the flow of control so that if we get an error
installing the converter, we don't try and remove it.