pdfwrite - avoid an error rendering a shading to an image
Following on from commit e57b1ffd95e189fedc68b2e235967dec649c20c3 which
corrected error handling when the image converter setup fails.

We'd prefer not to get an error at all, because that flags worrying
messages to the user. So here, if sx or sy are less than zero, which I
think indicates the shading is off the media, simply ignore the shading
altogether.