Bug 707667: Handle TTF with only platform 0 cmap tables
PDF requires embedded TrueType fonts to use the 3,1 (Windows Unicode), 3,0
(Windows Symbol) or 1,0 (Macintosh) cmap table.

The font in this files contains only one cmap table, and it is a 0,1 (Unicode
1.1 semantics - deprecated) cmap.

Technically invalid, but other consumers accept it.

Since it is Unicode, treat it the same as the 3,1 cmap (may not work as expected
in all cases).