PostScript startup code - Fix some missing newlines
Bug #707680 "missing newline character in "Last OS error" error message"

Commit baa1d0bbff203e2f22eac91925a52e0ce4d17b74 replaced the use of the
'=' PostScript function with the internal '=only' operator, because the
language defined function can be overridden by PostScript and could be
abused.

However, the two are not quite identical, the default implementation of
'=' includes a newline, whereas '=only' does not. This left a couple of
places failing to emit a newline.