Further improve the ZUGFeRD helper program
Bug #707694 "lib/zugferd.ps doesn't fill Length and ModDate"

Following further feedback from Thorsten Engel in the bug report, add
a /Size key to the Params sub-dictionary of the XML stream and find
it's value from the size of the file. Add an option to supply the
ModDate value for the Params sub-dictionary instead of using a dummy
value.