PDF interpreter - Remove spurious Matte handling from SMask groups
Bug #707689 "Stack-based buffer overflow in pdfi_trans_set_mask"

The crash occurs in code which should not be executed. It is processing
a /Matte entry in an ExtGstate /SMask entry, and Matte key/value pairs
are not valid in SMask groups, only for SMask images.

I've checked the image handling code and it correctly limits the array
handling to GS_MAX_COMPONENTS, returning a rangecheck error if the
number of components exceeds that.

This commit simply removes the offending code.