PDF interpreter - limit BC values
Bug #707690 "Stack-based buffer overflow in gs_begin_transparency_mask"

The interpreter wasn't limiting the number of components sent to the
graphics library, even though it did limit the number of values read.

Limit the value passed in params, and also throw an error if the number
of components exceeds our compile-time maximum. Discard the BC array
if that happens.

In addition; we cannot use the /BC array if there is no ColorSpace in
the associated Group attributes dictionary. In that case simply discard
the /BC array and give the user an error.