PDF interpreter - fix type 5 halftones
Bug #707631 "Problems with PreserveHalftoneInfo"

The problem is not with pdfwrite preserving halftones, but with the PDF
interpreter executing type 5 halftones.

This commit is based on the patch supplied by Alex in the bug report;
unfortunately that simple fix, while it does resolve the crash, leads to
many regressions in the test suite. It also deliberately doesn't
address the fact that the return value from gs_sethalftone_prepare() was
unchecked in several places.

Here we adopt Alex's patch, add the checks on the return value from
gs_sethalftone_prepare, but also correctly set up the identity transfer
function initially. The regressions were caused by the fact that the
default implementation essentially assumed a transfer function existed
and could map colour values incorrectly if one didn't.