Devices - use proper function to determine if SAFER is active
No bug for this one. Robin noticed on my original commit to prevent
device switching with SAFER that I was probing the context directly
when there is a function to return the SAFER status that we should use
instead.

There turn out to be several other places where we were already doing
the same thing (I had copied the code from elsewhere...), so this commit
alters those to use the function instead.