Rejig tags slightly.
We have a device structure member, graphics_type_tag, that we use to
hold the current tag. We use a bit in this indicate that the device
actually uses tags at all. This bit is never actually written to the
tags plane.

Given we only ever use 8 or 16 bits for the tag plane, why do we then
waste a bit using one of those for GS_DEVICE_ENCODES_TAGS ?

Here, we move GS_DEVICE_ENCODES_TAGS up to the 31st bit. This frees
another bit lower down for use with real tag values.

Also, we introduce an inline function, device_current_tag, to extract
the current tag. By using this, together with the existing function,
device_encodes_tags, we can restrict the use of GS_DEVICE_ENCODES_TAGS
within the broader code.