Make num_components/depth etc consistent w.r.t. tags.
When using tags, it seems we sometimes include the tag in the
number of components defined within a device, and sometimes
don't.

This causes confusion. Accordingly, move to consistently
always including the tag plane in num_components, and the tag
size in depth.

This has meant various changes within pdf14 and similar, but
the code at the end makes more sense to me.

The number of colorants is thus always num_components - has_tags
(where has_tags = device_encodes_tags(dev)).

This means we need to recalculate shifts and comps on colorspace
changes.