pdfwrite - fix negative Form BBox to clip calculation
Bug #707718 "Forms with BBox entirely negative are clipped incorrectly."

As the bug says. The problem is caused because the calculation didn't
originally anticipate both x or both y co-ordinates being less than
zero. This meant that we could incorrectly use the less negative value
as the minimum x or y for the clip, causing content to vanish.