Tweak byte ordering in output from bitrgbtags device.
Internally, the bitrgbtags device 'encode_color's as 0xTTRRGGBB
(which is sane, as the bitrgb device in 8 bit mode encodes them
as 0xRRGGBB).

This means that gs renders these into 32bit chunky memory devices.
Being gs, this means that these are in big endian format, meaning
that the order in memory is TT RR GG BB.

This is NOT the order we'd expect in the output (bmpcmp gets this
wrong, for one). Given that we are outputting using a modified
PAM format with a TUPL of RGB_TAG, we'd expect RR GG BB TT.

Accordingly, we reorder the bytes before writing here. We don't
believe these devices are actually being used by anyone other than
possibly for testing, not least because they are completely
undocumented.