Update bmpcmp to expect RGBT as RGBT rather than TRGB.
In the recent commit in which I changed bitrgbtags to produce
RGBT rather than TRGB, I stated that bmpcmp had expected it
this way. Turns out I was wrong, and bmpcmp was correctly looking
for TRGB.

BUT... RGBT makes much more sense, so update bmpcmp to match the
new bitrgbtags rather than reverting.