Update clist to properly record tags.
We have a device procedure "set_graphics_type_tag" that is used
to set the graphics type tag before particular drawing operations
start. The clist does not currently trap this, relying instead on
the clist playback code to set the tag when it starts operations.

This works fine for stuff like text or image operations, which
have defined starts, but works poorly for other operations, such as
shadings.

When we start a shading operation, the graphics library will call
set_graphics_type_tag (say, to change the tag from whatever it was
before (perhaps an image tag) to be a vector tag). The clist ignores
this.

The next actual graphical operation performed is then a
fill_trapezoid, which is written to the clist. When this is played
back, it therefore draws those pixels with the wrong tag.

The fix, implemented here, is to have the clist trap the
set_graphics_type_tag call and to write it into the clist.

The only complexity here is that when the clist is played back,
the particular arrangement of devices means that the same routine
is called again. We therefore need a way to identify whether the
clist is in reading or writing mode, and to only attempt to write
the command if we're in writing mode.

Strangely, it doesn't currently appear to be possible to easily
make this determination, so we update the clist code with a
'reading' flag that is kept up to date throughout device state
changes.