Add new clist render threading mechanism.
Credit is due to Biró Zoltán for the idea behind this, and the
bulk of the implementation.

While no current Ghostscript devices use this at the moment, it
makes sense to roll this into the released code to avoid him
having to pull it in again on every release.

Essentially, this commit adds a new mechanism for the operation
of multiple threaded rendering.

The existing mechanism makes it very simple for devices to use
multi-threaded rendering, while ensuring that they still get
data to their 'output' function in the expected order. The
downside to this mechanism is that once threads complete the
rendering task they can be blocked for a non-trivial amount of
time until it is their turn to output. This can be particularly
extreme if bands take different amounts of time to render
(perhaps some bands are much simpler than others).

The new mechanism vastly reduces the time where threads are
blocked, at the expense of data no longer reaching the 'output'
in the expected order. For some formats (say, writing to memory,
or writing fixed length data blocks to files at predicatable
offsets), this is not an issue.

We also update the downscaler to cope with being run in this
new way.