Bug 707726: "Fix" MacRomanEncoding reverse lookup
In a PDF with a TTF using a 1,0 cmap table, we are required to use an inverse
MacRomanEncoding to map from a glyph name back to a character code (to use
as input to the 1,0 cmap table).

The problem is that MacRomanEncoding has two entries that contain the "space"
glyph name: the "conventional" index 32, and also index 202. Using a binary
search to speed lookups, we're finding index 202 which is (almost) always
wrong.

So change the reverse lookup table to return 32 for both instances of the
"space" glyph name.