pdfwrite - when assessing text direction, account for descendants of CIDFonts
No bug report, this is from a custoemr email.

Commit e4ea7e1bd9e29813e273a4ca6d001b4279e8f17c stopped us from counting
clipped text as constributing towards text direction calculations. To do
so part of the fix was to move the calculation of the text direction
fomr the beginning of the text processing (pdf_text_begin) to the end of
text processing (pdf_text_release).

This all worked well but a non-obvious problem is exhibited by the
customer file; the text direction is calculated using the current font's
matrix, but for CIDFonts the current font at the start of text
processing may not be the same as at the end, because if we switch to a
descendant font then that becomes current.

The CIDfont and descendant font have separate font matrices and should
be combined. We have never done that which has always been incorrect.
Because of the commit, however, we were only cosidering the descendant
font, which altered the direction for some text.

In this commit we combine the CIDFont and descendant font matrices
before applying the CTM and use the combination to determine the text
direction.