Update MSVC makefile logic.
First off, add a new nmake version to the growing list. I wish
there was a nicer way of doing this.

Guessing the version of MSVC in use from the version of nmake
is a dodgy thing to do these days, when the toolset in use is
now separate from the IDE version.

Accordingly, we make some changes to the logic. We now allow
MSVC_VERSION to be set by the caller, and only set it with
our guess if it's not there already.

We do the same thing with MS_TOOLSET_VERSION.

We also make use of some environment variables that recent
versions of VS set so we can know the VS version exactly
if we are called from the solution (as hopefully most people
are doing) or from the command line prompt.

Rather than repeating long error prone paths, set them as a
variable and use that variable. This also aids cutting and
pasting.

Detect the variant (Community or Professional) from the
environment if possible, and just guess by checking what's
there if this fails.

Also, allow ourselves to look for "BuildTools" as an additional
variant in an attempt to solve bug 707435.