Bug 706445: Fix valgrind warnings in PDFI Tj handling.
We are ignoring error codes in the returns from functions here.
Not only is the initial point not set (which causes a warning when
we try to reset it at the end), but at least one of the other
calls returns with an error.

Here, we add appropriate error handling. And we do the same in
the TJ handling too, as it is very similar.