Bug 707753: Fix mismatch in pdf14 push/pops.
When we hit a text showing operation, this can cause a transparency
group to be pushed. Normally, this is popped again when we hit the
'ET' operator.

If we get a text showing operation outside of a BT/ET pair, then
we warn about it, but honour it anyway. If no ET is subsequently
encountered though, we never pop the transparency group.

The fix, implemented here, is to do the ET processing after the
text showing operation if we weren't in a BT to start with.