Fix some Coverity issues in new threading code.
CID 425267: Ensure that if we get errors in the threads we pass
this out as a failure.

CID 425268: Don't check 'options' for NULLity when we've already
dereferenced through it.

CID 425269: Avoid potential overflow in file offset calculation
if bands are really large.

Also, fix a potential overflow in the overprint code:

CID 424692: If the index of the tag plane is larger than
sizeof(int), we might have got overprint wrong. Unlikely, as
the tag plane needs to be at least 8 bits, but doesn't hurt to
patch it here.