oss-fuzz 68051: Clean up pattern and transparency device on error
Creating a pattern accumulator for a pattern which includes transparency ops,
if configuring the transparency compositor device triggered an error, neither
the compositor nor the accumulator device were being freed, leaving potential
problems either during a restore or garbage collection.

This also changes the accumulator device to be allocated in immovable memory,
since devices are supposed to be allocated as such.