Bug 705993: Fix SMask rendering problem.
When we apply an SMask, we need to change the profiles in
use. Instead of using 'default' ones (defaultgray, defaultrgb,
defaultcmyk), we need to use the 'ps' ones (psgray, psrgb,
pscmyk).

(Presumably this is to do with linearity? I don't know the logic
here, I just know that this is what the code seems to need to
do).

Accordingly, when an SMask is put into effect, the pdf14
compositor passes through the current graphic state, replacing
any 'default' profile, with the equivalent 'ps' one. This
operation is prompted by a call to gs_begin_transparency_mask.

Unfortunately, as this bug demonstrates, if the current color
space differs from the intended group colorspace when the SMask
is instantiated, the wrong thing happens. On the call to
gs_begin_transparency_mask, the replacement happens on the
current color space (which in this example is devicergb, because
we've just done "0 0 0 rg"). This space is rewritten as required.
But then we install the group colorspace, which swaps in
devicegray, which has not had the replacement done.

The fix for this is for us to ensure that the colorspace for
the group is in place BEFORE the call that triggers the
colorspace substitutions.

I have tried various ways to do this.

Firstly, I tried saving/restoring the gstate around the
begin_mask/execgroup/end_mask call, preparatory to
moving the group gstate setting out of the execgroup call.
This save/restore was enough to break lots of files in our
test repository.

Secondly, I tried just moving the setting of the groupgstate
out of the execgroup call to be before the begin_mask. This
gave diffs in the cluster results. Presumably because we're
setting some things that should not be set ?

Finally, the version that worked, I record the gstate before
I fiddle with it (as saved_gs). I set just the colorspaces
and colors (both filling and stroking). I then call
begin_mask/execgroup/end_mask as before. Then I restore the
colors and colorspaces from the saved version.