pdfwrite - update PDF/A-3 production for tech note 0010
Bug #707725 "PDF/A embedding files not pass VeraPDF validation"

As described in the PDF association technical note 0010, page 28,
"Associated files" it was not clear in the PDF/A-3 specification
(ISO 19005-3, Annex E) that the /AF key is *required* to be present in
at least one object for every embedded file. The PDF 2.0 specification
notes this as an optional key.

This commit adds an /AF array to the Catalog dictionary and lists every
embedded file there. This may not be 'correct' per the original document
but it passes the PDF/A-3 validation in both VeraPDF and Acrobat.

We could drop the Names reference (which apparently *is* optional), but
I'm choosing to leave it in place.

Additional testing with Acrobat showed that it expects the entries in
the /AF array to be indirect objects, even theough there is no apparent
requirement for this in either the PDF 2 or PDF/A-3 specifications. So
this commit also moves the Embedded File specifications to be objects in
their own right and references them via indirect references in both the
AF and Names objects.

This results in the supplied file passing, but I do not have any good
way to test this. I've run some tests on files that I used when I
originally wrote the embedded file support and everything appears to
work as expected, our regression testing shows no problems, but it
does not test this and cannot validate PDF/A compatibility anyway.