mswinpr2 device - suppress cancel dialog when no instance is available
Bug 701020 "ghostxps fails with mswinpr2 device"

The problem is that, while outputting the bitmap to the printer, the
mswinpr2 device tries to update the progress dialog (unless NoCancel
is set). If there is no instance (it is NULL) then we can't create the
dialog, which results in us failing to update it as we print, and
throwing an error.

While Ghostscript uses LoadLibrary and defines an entry point in the
DLL which allows it to capture the instance, the other interpreters
do not use explicit library loading and the DLLs do not permit the
instance to be captured. So every interpreter except Ghostscript fails
with the mswinpr2 device.

In this commit; if the instance is NULL then set nocancel to true, which
suppresses the cancel/progress dialog, avoiding the error.

It would be 'nicer' to get the instance but that ran into a nightmare
of dependencies on windows.h which refused to compile. This was simpler.