XPS interpreter - fix JPEG CMYK images and EXIF/APP13 resolution
Bug #704923 "When converting from XPS files with embedded images to PDF by Ghostscript (or MuPDF), the PDF result has color image incorrect."

As noted by Tor, the JPEG handling had two problems; firstly the image
samples for a CMYK image need to be inverted and secondly the code was
not taking account of either the APP13 marker resolution, nor the EXIF
resolution.

This commit adds an 'invert_decode' flag to the XPS image structure. If
we get a CMYK JPEG we set it to true, for all other images/ colour
spaces we set it to false. When building the graphcis library image
parameters structure, if the flag is true we invert the Decode array.

MuPDF already had code to read the APP13 and EXIF markers, and extract
the resolution from them, so I simply borrowed the code verbatim. I did
remove the processing of 'orientation' from the EXIF code since we don't
use it.

This fixes the bugs in the report, shows progressions in some seven of
the files in the test suite and completes the enhancement for bug