Bug 705284: Rejig gsaving around Image SMasks in the PDF interpreter.
When drawing an image with an SMask, we basically follow the
following steps:

  1. Draw the SMask.
  2. Draw the image through the SMask
  3. Pop the SMask

This has problems when step 2 of that process uses some
information in the current graphics state; we have seen
issues here with the color (and color_space) that is set
before step 1.

The issue is that the drawing of the SMask can corrupt the
current graphics state, thus affecting the graphical
operations that follow the SMask construction.

The obvious fix would be to amend the operations as follows:

  1. gsave
  2. Draw the SMask.
  3. grestore
  4. Draw the image through the SMask
  5. Pop the SMask

The problem with this is that the way we have stuff set up,
information about when to push/pull transparency stack is
stored as a flag in the graphics state too. Accordingly, if
we grestore, and lose the flag, the transparency handling
gets confused about when to remove the SMask.

The PDF interpreter rolls the logic for pushing and pulling
gstates while keeping the transparency state in sync into
its routines for handling q and Q. We can reuse those
functions in our code.

The solution, implemented here, is to move this flag up one
level. So:

  1. pdfi_op_q (gsave and push transparency)
  2. gsave
  3. Draw the SMask.
  4. Move the flag up one level.
  5. grestore (so the SMask is left in place).
  6. Draw the image through the SMask
  7. pdfi_op_Q (pop the transparency which pops the SMask and grestore)

This is a fairly clean solution, and avoids us needing to save
various small part of the gstate by ad-hoc copying.

We'd like to apply this kind of solution to other uses of the
SMask within the interpreter, but those aren't as nicely
encapsulated as this; in the other places the SMask usage does
not nicely match up with gsave/grestores. So we will continue
to think about that.