PDF interpreter - add a clip matching the selected Box in all cases
No bug report, the report is from a curstomer.

The customer was relying on the 'UseCropBox' switch applying a clip
rectangle to the output. However UseCropBox is only actually intended
to be used to set the media size (in place of MediaBox) so in general
we do not apply a clip (there is no point in applying a clip the size
of the media, we already clip to that).

When UseCropBox was not specified, the existing code used the MediaBox
but, crucially, because there is a CropBox in the PDF file it applied
the CropBox as a clip. This is a specific and I think documented case.

So it seemed like everything was working, except for the Use CropBox.

It seems we did do apply a clip in the past, so this commit applies a
rectangular clip which matches the selected Box, **unless** the selected
box is the MediaBox *and* there is a CropBox in the PDF file. In that
case we always apply a clip from the CropBox.

This threw up a bug in the xpswrite device. There was a path where the
xpswrite device emitted a clip without setting the device's
'clip_writtten' flag. This could lead to us writing a path with two
'Clip' attributes. Interestingly the Microsoft XPS viewer doesn't care
about that, but Expat throws an error.

Fixed by setting the relevant flag.