Bug 707761: Fix Object moving according to resolution.
Some of the object rendering code was looking at the penum->dxy
value to decide how to alter y.

This value is calculated from a double value by converting to
float, with a small epsilon value added for rounding.

In this example, the double value was very small and negative,
and as the resolution ticks over to 410, the net effect of the
rounding was to make it go to zero, causing the logic to change.

Here, we fix the code to look at the the double value rather
than the fixed version, which solves the bug.