Bug 707767: Add alternative put_image method for mem_planar devices.
The put_image device proc is always called with contone (byte, or
16bit) data.

The pdf14 implementation of this (which everything goes through)
tries to pass the data onto the target device. First, it tries
to pass the data with the alpha intact - most devices will
refuse this at the moment.

If refused, the pdf14 implementation blends the alpha, and then
tries to pass it again. Again, the device can either accept this
or refuse it.

If refused, the pdf14 implementation resends the data as an image
in the appropriate colorspace for the device.

Unfortunately, this last step does not work with separation
supporting devices, in that any separation planes are dropped.
The code notes that all separation supporting devices should
accept the data as a put_image call.

So, tiffsep1 is implemented as a mem_planar device, and the
mem_planar device currently only has a put_image proc that can
cope when the device is a contone one.

The solution, implemented here, is to add a new put_image
implementation to mem_planar that can be used for non-contone
devices. This turns the incoming data into fill_rectangle_hl_color
calls. We don't have a pgs to pass through, but we get away with
this because the place it ends up doesn't actually need it.