Allow for Additive spot color devices.
Currently, Ghostscript assumes that devices that want spot
colors will be subtractive (i.e. CMYK).

It would be nice to support devices that want RGB+Spots.

The logic that needs changing for this is duplicated in
2 places, check_DeviceN_component_names and
check_Separation_component_names.

Currently, we force a fallback to blending the spot (by
forcing use_alt_cspace) unless we're a devn supporting
device that happens to be a pdf14 device. This is to
ensure that we handle transparency spaces that happen
to be RGB + spots correctly according to the pdf spec.

We modify this logic here so that we also don't drop back
if we are devn supporting device that specifically claims
to want spots to be maintained. We allow the device to make
such a claim by adding a new gxdso.

It's possible that we should just say "all devn additive
space devices should cope with spots", but I lack the bravery
to make that change. By gating the behaviour change on a
gxdso we can ensure that we limit any bad effects to specific
devices.

There is one nasty aspect here. When mapping a separation to
a colorspace, we have previously always been able to assume
that all the components are subtractive (both process and
spots).

Now, for additive spaces, we need to invert the process
color values, but in cmap_separation_direct, we have no way of
knowing how many process colorants there are. Accordingly,
we call our gxdso and arrange for that to return the number of
additive process colors we have.