Fix the SEGVs seen when psdrgb runs in clist mode.
Now psdrgb has num_components > 3, the ICC code helpfully
sets the default profile to be a CMYK one. This causes
the clist code to think that we should have 4 components
for some groups leading to reading/writing code gettting
out of step.

This fixes that issue, but testing shows that psdrgb is
losing spots in the clist case. I will investigate that.