Fix psdrgb fillpage issue.
The following command:

 gs -sDEVICE=psdrgb -o out.psd SpotColorTest3.pdf

results in the background, which is supposed to be white,
being filled with a white+random spot values.

SpotColorTest3 is an RGB+3 spots file.

This is because the initial fillpage maps rgb white and
the rgb_cs_to_psdrgb_cm routine fails to set the spots to
zero.

This code is looking at devn_params.separations.num_separations
but this appears to be 0. Tracing through the equivalent
code used for psdcmyk, that is looking at num_components instead,
which seems a better source to me. Update the rgb routines to use
the same thing too.