Fix asserts seen in overnight testing with debug-Z@.
The following command would assert:

 debugbin/gswin64c.exe -ZL -sDEVICE=psdrgb -r72 -dMaxBitmap=1000
 -dBandHeight=2000 -o out%d.psd tests/pdf/cff_extra_hdr_size.pdf

This is due to us arriving in overprint_fill_rectangle_hl_color
without having set op_state to tell us whether we were doing
a fill or a stroke.

This is because the overprint device is pushed for all
separation supporting devices, but the state was only being
updated for CMYK or DevN ones.

It seems that we need to enable overprint for at least pdf14
RGB devices in at least some cases, so it's easiest to just
drop the colorspace requirement when updating the op_state.